pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nibbler-bot"
version = "0.1.0"
description = "A Telegram calorie tracker powered by OpenAI vision."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0,<1.0.0",
  "openai>=1.75.0,<2.0.0",
  "pillow>=11.1.0,<12.0.0",
  "python-telegram-bot[job-queue]>=22.0,<23.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5,<9.0.0",
]

[project.scripts]
nibbler-bot = "nibbler_bot.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
